refactor(csi_volume): extract shared helpers into dedicated helpers.go
